requirements.txt
line-bot-sdk
flask
cx_Oracle
gunicorn
python-dotenv
